requirements.txt
aiohttp==3.7.4
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
discord.py==1.6.0
distlib==0.3.1
Django==3.1.5
filelock==3.0.12
idna==3.1
Jinja2==2.11.2
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mpld3==0.5.2
multidict==5.1.0
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
pygame==2.0.1
pygame-menu==3.5.5
pyparsing==2.4.7
pyperclip==1.8.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
stego-lsb==1.3.1
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
virtualenv==20.2.2
yarl==1.6.3
